# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Legacy Java Modernization Agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type AnalysisStatus =
  (typeof AnalysisStatus)[keyof typeof AnalysisStatus];

export const AnalysisStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
  failed: "failed",
} as const;

export interface Analysis {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  status: AnalysisStatus;
  /** @nullable */
  currentStep?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateAnalysisBody {
  name: string;
  /** @nullable */
  description?: string | null;
}

export interface Repository {
  id: number;
  analysisId: number;
  name: string;
  javaCode: string;
  /** @nullable */
  packageStructure?: string | null;
  createdAt: string;
}

export interface CreateRepositoryBody {
  name: string;
  javaCode: string;
  /** @nullable */
  packageStructure?: string | null;
}

export interface AnalysisResult {
  id: number;
  analysisId: number;
  step: number;
  stepName: string;
  content: string;
  createdAt: string;
}