# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Legacy Java Modernization Agent API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all analysis sessions
 */
export const ListAnalysesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  status: zod.enum(["pending", "in_progress", "completed", "failed"]),
  currentStep: zod.number().nullish(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});
export const ListAnalysesResponse = zod.array(ListAnalysesResponseItem);

/**
 * @summary Create a new analysis session
 */
export const CreateAnalysisBody = zod.object({
  name: zod.string(),
  description: zod.string().nullish(),
});

/**
 * @summary Get an analysis session
 */
export const GetAnalysisParams = zod.object({
  id: zod.coerce.number(),
});

export const GetAnalysisResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  status: zod.enum(["pending", "in_progress", "completed", "failed"]),
  currentStep: zod.number().nullish(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Delete an analysis session
 */
export const DeleteAnalysisParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List repositories in an analysis
 */
export const ListRepositoriesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListRepositoriesResponseItem = zod.object({
  id: zod.number(),
  analysisId: zod.number(),
  name: zod.string(),
  javaCode: zod.string(),
  packageStructure: zod.string().nullish(),
  createdAt: zod.date(),
});
export const ListRepositoriesResponse = zod.array(ListRepositoriesResponseItem);

/**
 * @summary Add a repository to an analysis
 */
export const CreateRepositoryParams = zod.object({
  id: zod.coerce.number(),
});

export const CreateRepositoryBody = zod.object({
  name: zod.string(),
  javaCode: zod.string(),
  packageStructure: zod.string().nullish(),
});

/**
 * @summary Remove a repository from an analysis
 */
export const DeleteRepositoryParams = zod.object({
  id: zod.coerce.number(),
  repoId: zod.coerce.number(),
});

/**
 * @summary List all generated results for an analysis
 */
export const ListResultsParams = zod.object({
  id: zod.coerce.number(),
});

export const ListResultsResponseItem = zod.object({
  id: zod.number(),
  analysisId: zod.number(),
  step: zod.number(),
  stepName: zod.string(),
  content: zod.string(),
  createdAt: zod.date(),
});
export const ListResultsResponse = zod.array(ListResultsResponseItem);

/**
 * @summary Run a specific analysis step (SSE streaming)
 */
export const runAnalysisStepPathStepMax = 6;

export const RunAnalysisStepParams = zod.object({
  id: zod.coerce.number(),
  step: zod.coerce.number().min(1).max(runAnalysisStepPathStepMax),
});

/**
 * @summary Run all 6 analysis steps sequentially (SSE streaming)
 */
export const RunFullAnalysisParams = zod.object({
  id: zod.coerce.number(),
});