# lib/api-zod/src/generated/types/analysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Legacy Java Modernization Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisStatus } from "./analysisStatus";

export interface Analysis {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  status: AnalysisStatus;
  /** @nullable */
  currentStep?: number | null;
  createdAt: Date;
  updatedAt: Date;
}