# lib/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Legacy Java Modernization Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalysisResult {
  id: number;
  analysisId: number;
  step: number;
  stepName: string;
  content: string;
  createdAt: Date;
}