# lib/api-zod/src/generated/types/createRepositoryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Legacy Java Modernization Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRepositoryBody {
  name: string;
  javaCode: string;
  /** @nullable */
  packageStructure?: string | null;
}